Decimals in every numeric field; drop the "All sites" view
The 0-1 error was the browser, not our validation: an integer column meant
`<input type="number">` with the default step of 1, so Chrome refused a decimal
with "the two nearest valid values are 0 and 1" before the form was submitted.
0019 fixed the score columns only, leaving the tallies, Ahrefs metrics, backlink
count, search volumes and every sort order still whole.

- 0020: those columns become numeric(_,2). rankings.position (a rank, 1..100,
  importer-written) and qa_*_audit.sheet_row stay integer — neither has an input.
- ranking_grid / ranking_grid_multi declared sort_order as integer, so they had
  to be recreated returning numeric or the grid would fail at call time.
- Parsers: build-row and the volume grid use parseDecimalField; the backlink
  count no longer goes through parseInt. Every number input carries step="0.01".
- SiteSelector drops "All sites": the report is always one site, and a missing
  or unknown ?site= resolves to the first (lib/sites.ts), so the five pages that
  used to show every site now match what the selector says.

+++ src/lib/manage/build-row.ts
@@ -1,5 +1,5 @@
 import type { Field } from "./entities";
-import { parseIntegerField } from "@/lib/numeric";
+import { parseDecimalField } from "@/lib/numeric";
 
 type Raw = Record<string, string | undefined>;
 export type BuiltRow = { data: Record<string, string | number | boolean | null>; errors: string[] };
@@ -32,7 +32,9 @@ export function buildRow(fields: Field[], raw: Raw): BuiltRow {
             data[field.name] = typeof field.defaultValue === "number" ? field.defaultValue : null;
           }
         } else {
-          const r = parseIntegerField(trimmed, field.label);
+          // Every numeric column behind these forms is numeric(_,2) (sort_order,
+          // search volume), so decimals are stored as typed instead of rejected.
+          const r = parseDecimalField(trimmed, field.label);
           if (!r.ok) errors.push(r.error);
           else data[field.name] = r.value;
         }
